chore(functional tests): extract and activate webpack watch in debug server
very handy to understand what's going on, previously it was not
recompiling the bundle and would lead to WTF + F5 F5 F5
