+++ scripts/dev-functional-tests-debug-server.js
@@ -1,4 +1,7 @@
 import startServer from '../functional-tests/startServer.js';
+import watch from './dev-functional-tests-compile-watch.js';
+
+watch(() => {});
 
 startServer()
   .then(serverInstance => console.log('http://localhost:' + serverInstance.address().port))
